Added additional URDFs
URDFs of just the a1 robot and the mobile px100 robot have been added.
Different variations of the Legged Manipulator robot have been updated with adjusted file paths. CAUTION: with the adjusted file paths, if this repository is downloaded the file paths to the meshes will need to be updated.