Implementing Cartesian Trunk task.
To track the trunk of the legged manipulator, a Cartesian task has been added for the trunk. This allows the trunk try and maintain a defined position within the world frame. Furthermore, it allows for better tracking of the CoM as the jacobian and position of the trunk is now defined.