Moving weight variables and changing orientation of EE reference frames.
For easy adjustment of cartesian task weights, they are now initialised in the __init__ section of the class.

The jacobain of the EE is now taken in the local frame which is aligned with the world frame. This ensures that the orientation of all frames is initally the same, making it easier to control.