Developing the wrapper to make it modular.
So that this wrapper can be used with any legged manipulator, the wrapper has been developed to find the joint and frame ids dynamically. The initial posture is also modular as the initial joint configuration is found through solving a set of QP problems to the point where the legged manipulator is in a stable configuration.