Further improving robot and QP wrappers modulatrity.
Several bugs have been fixed where URDF specific parameters were not accounted for, such as the number of configuration and velocity dimensions. The names of the end efffector frames are now passed into the robot wrapper as a list to streamline the wrapper.

For the QP wrapper, the number of solutions can now be specified to improve the modulatiry of the code.

sim2 has been developed to account for these changes and test different URDFs.