Integrate bug fix, applying IMU and changing vel calc methods.
Major bug fix related to do with the integrate function only returning the joint positions per unit time. The input velocity is now multiplied by the change in time to fix this bug.

The initial configuration now takes into account the joint limits of the robot through adjusting the neutral configuration and applying more advanced velocity calculation methods.

In the simulation the robot now takes the IMU data as an input. More work is required to make this data usable.